+++ packages/@valapi/web-client/src/service/DailyTicket.ts
@@ -1,0 +1,48 @@
+import type { AxiosResponse } from "axios";
+
+import { WebClientService } from "../client/WebClientService";
+
+export namespace DailyTicket {
+    // response
+
+    export interface DailyTicket {
+        Version: number;
+        DailyRewards: {
+            RemainingLifetimeSeconds: number;
+            BonusMilestonesPending: number;
+            Milestones: Array<{
+                Progress: number;
+                BonusApplied: boolean;
+            }>;
+        };
+        ProcessedMatches: Array<{
+            ID: string;
+            ProgressBefore: number;
+            ProgressAfter: number;
+            XP: number;
+            SoftCurrency: number;
+            WasPenalized: boolean;
+            BonusesApplied: number;
+            DailyBonusState: [boolean, boolean, boolean, boolean];
+            RewardGrants: any; // * unknown
+        }>;
+    }
+}
+
+export class DailyTicket extends WebClientService {
+    /**
+     * @param {string} subject Player UUID
+     * @returns {Promise<AxiosResponse<DailyTicket.DailyTicket>>}
+     */
+    public async get(subject: string): Promise<AxiosResponse<DailyTicket.DailyTicket>> {
+        return await this.axios.get(`${this.apiRegion.url.playerData}/daily-ticket/v1/${subject}`);
+    }
+
+    /**
+     * @param {string} subject Player UUID
+     * @returns {Promise<AxiosResponse<DailyTicket.DailyTicket>>}
+     */
+    public async renew(subject: string): Promise<AxiosResponse<DailyTicket.DailyTicket>> {
+        return await this.axios.post(`${this.apiRegion.url.playerData}/daily-ticket/v1/${subject}/renew`);
+    }
+}